Added the fox Texture, script, sight script, and the fixed toggle script
When we integrate let me deal with the fox as I know how to work with
it. If you don't know how to use ANY of my scripts let me know so that I
can help as soon as I can. Other than that this is all I have to give
you.